refactor: split markdown output and preview (wip)
