+++ paper2llm-web/src/web/components/markdown-preview/index.ts
@@ -1,0 +1,3 @@
+// Re-export the main component
+export { default } from "../MarkdownPreview";
+export * from "./types";
